// src/utils/types.ts
export interface User {
  id?: string;
  email: string;
  firstName: string;
  lastName: string;
  password?: string;
  confirmPassword?: string;
  phoneNumber: string;
  profileImage?: string;
  role: UserRole;
  coverPhoto?: string;
  totalPurchases?: number;
  createdAt?: string;
}

export enum UserRole {
  customer = "customer",
  vendor = "vendor",
}